# docs/API_OPENAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BehavioralWell \u2014 Web + Android Shared Platform",
    "description": "Privacy-conscious multimodal behavioral risk detection, digital phenotyping, and early intervention API engine.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/risk/{user_id}/latest": {
      "get": {
        "tags": [
          "Risk Staging & Engine"
        ],
        "summary": "Get Current Risk",
        "operationId": "get_current_risk_api_v1_risk__user_id__latest_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskAssessmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/current": {
      "get": {
        "tags": [
          "Risk Staging & Engine"
        ],
        "summary": "Get Current Risk",
        "operationId": "get_current_risk_api_v1_risk_current_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "usr_demo12345",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskAssessmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/history": {
      "get": {
        "tags": [
          "Risk Staging & Engine"
        ],
        "summary": "Get Risk History",
        "operationId": "get_risk_history_api_v1_risk_history_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "usr_demo12345",
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/explanation": {
      "get": {
        "tags": [
          "Risk Staging & Engine"
        ],
        "summary": "Get Risk Explanation",
        "operationId": "get_risk_explanation_api_v1_risk_explanation_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "usr_demo12345",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/risk/{user_id}/latest": {
      "get": {
        "tags": [
          "Risk Staging & Engine"
        ],
        "summary": "Get Current Risk",
        "operationId": "get_current_risk_api_risk__user_id__latest_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskAssessmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/risk/current": {
      "get": {
        "tags": [
          "Risk Staging & Engine"
        ],
        "summary": "Get Current Risk",
        "operationId": "get_current_risk_api_risk_current_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "usr_demo12345",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskAssessmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/risk/history": {
      "get": {
        "tags": [
          "Risk Staging & Engine"
        ],
        "summary": "Get Risk History",
        "operationId": "get_risk_history_api_risk_history_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "usr_demo12345",
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/risk/explanation": {
      "get": {
        "tags": [
          "Risk Staging & Engine"
        ],
        "summary": "Get Risk Explanation",
        "operationId": "get_risk_explanation_api_risk_explanation_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "usr_demo12345",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContributorFactor": {
        "properties": {
          "feature": {
            "type": "string",
            "title": "Feature"
          },
          "modality": {
            "type": "string",
            "title": "Modality"
          },
          "direction": {
            "type": "string",
            "title": "Direction"
          },
          "z_score": {
            "type": "number",
            "title": "Z Score"
          },
          "human_explanation": {
            "type": "string",
            "title": "Human Explanation"
          }
        },
        "type": "object",
        "required": [
          "feature",
          "modality",
          "direction",
          "z_score",
          "human_explanation"
        ],
        "title": "ContributorFactor"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModalityBreakdown": {
        "properties": {
          "keyboard": {
            "type": "number",
            "title": "Keyboard"
          },
          "usage": {
            "type": "number",
            "title": "Usage"
          },
          "motion": {
            "type": "number",
            "title": "Motion"
          },
          "work": {
            "type": "number",
            "title": "Work"
          },
          "mobility": {
            "type": "number",
            "title": "Mobility"
          }
        },
        "type": "object",
        "required": [
          "keyboard",
          "usage",
          "motion",
          "work",
          "mobility"
        ],
        "title": "ModalityBreakdown"
      },
      "RiskAssessmentResponse": {
        "properties": {
          "risk_score": {
            "type": "number",
            "title": "Risk Score"
          },
          "stage": {
            "type": "integer",
            "title": "Stage"
          },
          "stage_label": {
            "type": "string",
            "title": "Stage Label"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "trend": {
            "type": "string",
            "title": "Trend"
          },
          "persistence_days": {
            "type": "integer",
            "title": "Persistence Days"
          },
          "top_contributors": {
            "items": {
              "$ref": "#/components/schemas/ContributorFactor"
            },
            "type": "array",
            "title": "Top Contributors"
          },
          "modality_scores": {
            "$ref": "#/components/schemas/ModalityBreakdown"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "risk_score",
          "stage",
          "stage_label",
          "confidence",
          "trend",
          "persistence_days",
          "top_contributors",
          "modality_scores",
          "timestamp"
        ],
        "title": "RiskAssessmentResponse"
      },
      "RiskHistoryResponse": {
        "properties": {
          "history": {
            "items": {
              "$ref": "#/components/schemas/RiskAssessmentResponse"
            },
            "type": "array",
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "history"
        ],
        "title": "RiskHistoryResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}